requirements.txt
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
alpaca-py>=0.10.0
matplotlib>=3.7.0
